tests/wpt/web-platform-tests/html/canvas/element/path-objects/2d.path.isPointInpath.multi.path.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.path.isPointInpath.multi.path</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.isPointInpath.multi.path</h1>
<p class="desc">Verify the winding rule in isPointInPath works for path object.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Verify the winding rule in isPointInPath works for path object.");
_addTest(function(canvas, ctx) {

canvas.width = 200;
canvas.height = 200;

// Testing default isPointInPath with Path object');
path = new Path2D();
path.rect(0, 0, 100, 100);
path.rect(25, 25, 50, 50);
_assertSame(ctx.isPointInPath(path, 50, 50), true, "ctx.isPointInPath(path, 50, 50)", "true");
_assertSame(ctx.isPointInPath(path, 50, 50, undefined), true, "ctx.isPointInPath(path, 50, 50, undefined)", "true");
_assertSame(ctx.isPointInPath(path, NaN, 50), false, "ctx.isPointInPath(path, NaN, 50)", "false");
_assertSame(ctx.isPointInPath(path, 50, NaN), false, "ctx.isPointInPath(path, 50, NaN)", "false");

// Testing nonzero isPointInPath with Path object');
path = new Path2D();
path.rect(0, 0, 100, 100);
path.rect(25, 25, 50, 50);
_assertSame(ctx.isPointInPath(path, 50, 50, 'nonzero'), true, "ctx.isPointInPath(path, 50, 50, 'nonzero')", "true");

// Testing evenodd isPointInPath with Path object');
path = new Path2D();
path.rect(0, 0, 100, 100);
path.rect(25, 25, 50, 50);
assert_false(ctx.isPointInPath(path, 50, 50, 'evenodd'));


});
</script>
